requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Flask-Migrate
Flask-WTF
# flask-bcrypt
cryptography
pymysql
email_validator